+++ sketch-core/src/commonTest/kotlin/com/github/panpf/sketch/core/common/test/http/HttpStackTest.kt
@@ -1,0 +1,4 @@
+package com.github.panpf.sketch.core.common.test.http
+
+class HttpStackTest {
+}