+++ sketch-http-core/src/commonTest/kotlin/com/github/panpf/sketch/http/core/common/test/http/HttpStackTest.kt
@@ -1,4 +1,0 @@
-package com.github.panpf.sketch.http.core.common.test.http
-
-class HttpStackTest {
-}